Remove vendor dependencies and cleanup scripts. (#11)
* Remove vendor dependencies and cleanup scripts.

SubQL latest release support all the fixes and enhancements that was made by @bryanchriswhite on vendor packages. So there is no need to keep them on the repo and overcomplicate the build process.

* Update README

* Add amd64 and arm64 build targets and precommit script

Ensure both architectures are properly built and tested for CI compatibility. Introduced a precommit hook to streamline the process and made necessary amendments to the Dockerfile for ARM64 support.

---------

Co-authored-by: Daniel Olshansky <olshansky.daniel@gmail.com>