[Feature] Implement Validators Handling (#36)
This PR introduces enhancements to the genesis block processing and adds
support for validator-related transactions.
Also, it modify tilt files to change graphql service to be exposed on
port 3001 instead of 3000 which is already use by indexer. This last
change is to been able to expose both to the hosts without conflicts.